#6 Moved BookUrlField to a separate file
