+++ lib/features/book_selection/ui/pages/book_selection_page.dart
@@ -1,7 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:storyscape/core/routing/routes.dart';
+import 'package:storyscape/features/book_selection/ui/pages/book_url_field.dart';
 
 class BookSelectionPage extends StatelessWidget {
   const BookSelectionPage({super.key});
@@ -19,28 +19,6 @@ class BookSelectionPage extends StatelessWidget {
             onFinished: (String url) => BookReadingRoute(url: url).push<void>(context),
           ),
         ),
-      ),
-    );
-  }
-}
-
-class BookUrlField extends HookWidget {
-  const BookUrlField({required this.onFinished, super.key});
-
-  final void Function(String) onFinished;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextEditingController controller = useTextEditingController();
-
-    return TextField(
-      controller: controller,
-      onEditingComplete: () => onFinished(controller.value.text),
-      keyboardType: TextInputType.url,
-      decoration: InputDecoration(
-        labelText: 'bookSelection.urlFieldLabel'.tr(),
-        suffixIcon:
-            IconButton(onPressed: () => onFinished(controller.value.text), icon: const Icon(Icons.arrow_forward)),
       ),
     );
   }
